[NWPS-1027] - Moved and renamed hee-anchorlinks directory to anchorlinks within new structure
