+++ app/assets/hee/blocks/content/sidebar/anchorlinks/_anchorlinks.scss
