+++ app/assets/hee/blocks/content/sidebar/anchorlinks/anchorlinks.js
