+++ app/assets/hee/blocks/content/sidebar/anchorlinks/toc.js
