+++ app/assets/hee/hee.scss
@@ -60,7 +60,7 @@
 @import '../components/header-search';
 @import '../components/hee-table-2';
 @import '../components/hee/hee-appliesto/appliesto';
-@import '../components/hee/hee-anchorlinks/anchorlinks';
+@import 'blocks/content/sidebar/anchorlinks/anchorlinks';
 @import 'blocks/scaffolding/notifications/phasebanner/phasebanner';
 @import '../components/hee/hee-tabs/tabs';
 @import '../components/nhsuk-hee/nhsuk-hee-tables/tables2'; //?
