+++ app/views/example.html
@@ -45,7 +45,7 @@
 
 {# To Sort #}
 {# HEE Components #}
-{%- from 'components/hee/hee-anchorlinks/macro.njk' import anchorLinks %}
+{%- from 'hee/blocks/content/sidebar/anchorlinks/macro.njk' import anchorLinks %}
 {%- from 'components/hee/hee-appliesto/macro.njk' import appliesTo %}
 {%- from 'components/hee/hee-media/macro.njk' import media %}
 {%- from 'components/hee/hee-navmap/macro.njk' import navMap %}
