+++ app/views/template.html
@@ -30,7 +30,7 @@
 {# CONTENT (BLOCKS) COMPONENTS #}
 
 {# HEE Components #}
-{%- from 'components/hee/hee-anchorlinks/macro.njk' import anchorLinks %}
+{%- from 'hee/blocks/content/sidebar/anchorlinks/macro.njk' import anchorLinks %}
 {%- from 'hee/blocks/content/main/card--author/macro.njk' import authorCard %}
 {%- from 'hee/blocks/content/sidebar/card--person/macro.njk' import personCard %}
 {%- from 'hee/blocks/content/sidebar/card--department/macro.njk' import departmentCard %}
